requirements.txt
datasets==2.12.0
langchain==0.0.169
matplotlib==3.7.1
numpy==1.24.3
Pillow==9.5.0
Requests==2.30.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
